requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
openai
tiktoken
llama_index
